Enable quality check plugins with functions
This avoids applying the plugin(s) if not explicitly required.
